Analysing Customer Sentiment from Text Using Microsoft Fabric AI functions
This notebook leverages Microsoft Fabric AI functions, which uses the capabilities of large language models (LLMs), to analyse customer satisfaction sentiment for a clothing store using feedback collected from customers. The goal is to evaluate monthly trends in customer sentiment, helping the business understand how customers feel about its products and services over time and helping the business to implement continuous improvement.

In this case, the ai.analyze_sentiment function is used to detect sentiment in text, identifying whether the emotional tone is positive, neutral, or negative. The result data can be written to a Lakehouse and visualised in a Power BI, providing insights for decision-makers.